chore: replace segment.rs with segment_new.rs
